Added displaying of old cached data as placeholders
